\name{expandFactorLevels}
\alias{expandFactorLevels}
\title{expandFactorLevels}
\description{expand a factor in all dataset to encompass levels of all sets}
\usage{expandFactorLevels(datasets, varName, .omitWarning = FALSE)}
\arguments{
  \item{datasets}{list of data.frames}
  \item{varName}{scalar string of variable holding the factor}
  \item{.omitWarning}{
}
}

\value{list of datasets with each entries column releveled}

\author{Thomas Wutzler}





